Big changes for this one.
Overhauled game.py file to be towards the style
of game I plan on going for. Created a small world
with a background and character for testing purposes.

Nothing set in stone, but this is the style I am
wanting. All images are going to change, but for
the purpose of learning and development I will
continue using these.